requirements.txt
aniso8601==4.1.0
asn1crypto==0.24.0
astroid==1.6.5
atomicwrites==1.3.0
attrs==18.2.0
autopep8==1.4.3
backports.functools-lru-cache==1.5
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
configparser==3.7.1
coverage==4.5.2
coveralls==1.5.1
cryptography==2.5
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
Flask-RESTful==0.3.7
funcsigs==1.0.2
futures==3.1.1
gunicorn==19.9.0
idna==2.8
ipaddress==1.0.22
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==5.0.0
nose==1.3.7
pathlib2==2.3.3
pep8==1.7.1
pluggy==0.8.1
psycopg2==2.7.7
psycopg2-binary==2.7.7
pur==5.2.1
py==1.7.0
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pylint==1.9.4
pyOpenSSL==19.0.0
pytest==4.2.0
pytest-cov==2.6.1
python-coveralls==2.9.1
python-dotenv==0.10.1
pytz==2018.9
requests==2.21.0
scandir==1.9.0
singledispatch==3.4.0.3
six==1.12.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.11.1
